Moved cache out of 'util' since we're gonna use it in other things * Also moved functions used by auth views into auth.core
